requirements.txt
apache-beam[gcp]==2.53.0
google-cloud-pubsub==2.21.0
